--- a/packages/ssz/test/lodestarTypes/phase0/viewDU/listValidator.ts
+++ b/packages/ssz/test/lodestarTypes/phase0/viewDU/listValidator.ts
@@ -40,6 +40,11 @@ for (let i = 0; i < PARALLEL_FACTOR; i++) {
 }
 const validatorRoot = new Uint8Array(32);
 
+/**
+ * Similar to ListCompositeTreeViewDU with some differences:
+ * - if called without params, it's from hashTreeRoot() api call, no need to compute root
+ * - otherwise it's from batchHashTreeRoot() call, compute validator roots in batch
+ */
 export class ListValidatorTreeViewDU extends ListCompositeTreeViewDU<ValidatorNodeStructType> {
   constructor(
     readonly type: ListCompositeType<ValidatorNodeStructType>,
@@ -69,8 +74,18 @@ export class ListValidatorTreeViewDU extends ListCompositeTreeViewDU<ValidatorNo
       number,
       ContainerNodeStructTreeViewDU<typeof ValidatorType>
     >;
-    const indicesChanged = Array.from(this.viewsChanged.keys()).sort((a, b) => a - b);
-    const endBatch = indicesChanged.length - (indicesChanged.length % PARALLEL_FACTOR);
+
+    const indicesChanged: number[] = [];
+    for (const [index, viewChanged] of viewsChanged) {
+      // should not have any params here in order not to compute root
+      viewChanged.commit();
+      // `validators.get(i)` was called but it may not modify any property, do not need to compute root
+      if (viewChanged.node.h0 === null) {
+        indicesChanged.push(index);
+      }
+    }
+    const sortedIndicesChanged = indicesChanged.sort((a, b) => a - b);
+    const endBatch = sortedIndicesChanged.length - (sortedIndicesChanged.length % PARALLEL_FACTOR);
     // nodesChanged is sorted by index
     const nodesChanged: {index: number; node: Node}[] = [];
     // commit every 16 validators in batch
@@ -80,7 +95,7 @@ export class ListValidatorTreeViewDU extends ListCompositeTreeViewDU<ValidatorNo
         batchLevel4Bytes.fill(0);
       }
       const indexInBatch = i % PARALLEL_FACTOR;
-      const viewIndex = indicesChanged[i];
+      const viewIndex = sortedIndicesChanged[i];
       const viewChanged = viewsChanged.get(viewIndex);
       if (viewChanged) {
         validatorToChunkBytes(level3ByteViewsArr[indexInBatch], level4BytesArr[indexInBatch], viewChanged.value);
@@ -96,12 +111,10 @@ export class ListValidatorTreeViewDU extends ListCompositeTreeViewDU<ValidatorNo
         digestNLevel(batchLevel3Bytes, 3);
         // commit all validators in this batch
         for (let j = PARALLEL_FACTOR - 1; j >= 0; j--) {
-          const viewIndex = indicesChanged[i - j];
+          const viewIndex = sortedIndicesChanged[i - j];
           const indexInBatch = (i - j) % PARALLEL_FACTOR;
           const viewChanged = viewsChanged.get(viewIndex);
           if (viewChanged) {
-            // should not have any params here in order not to compute root
-            viewChanged.commit();
             const branchNodeStruct = viewChanged.node;
             byteArrayIntoHashObject(validatorRoots[indexInBatch], 0, branchNodeStruct);
             nodesChanged.push({index: viewIndex, node: viewChanged.node});
@@ -114,12 +127,10 @@ export class ListValidatorTreeViewDU extends ListCompositeTreeViewDU<ValidatorNo
 
     // commit the remaining validators, we can do in batch too but don't want to create new Uint8Array views
     // it's not much different to commit one by one
-    for (let i = endBatch; i < indicesChanged.length; i++) {
-      const viewIndex = indicesChanged[i];
+    for (let i = endBatch; i < sortedIndicesChanged.length; i++) {
+      const viewIndex = sortedIndicesChanged[i];
       const viewChanged = viewsChanged.get(viewIndex);
       if (viewChanged) {
-        // commit
-        viewChanged.commit();
         // compute root for each validator
         viewChanged.type.hashTreeRootInto(viewChanged.value, validatorRoot, 0);
         byteArrayIntoHashObject(validatorRoot, 0, viewChanged.node);